+++ crates/full-node/sov-sequencer/src/lib.rs
@@ -20,6 +20,7 @@ pub use config::{SequencerConfig, SequencerKindConfig};
 pub use rest_api::SequencerApis;
 use serde::Serialize;
 use sov_modules_api::{RuntimeEventProcessor, RuntimeEventResponse};
+use sov_rollup_interface::common::SlotNumber;
 use sov_rollup_interface::TxHash;
 pub use tx_status::TxStatusManager;
 
@@ -33,26 +34,19 @@ pub struct SubmitBatchReceipt {
     pub tx_hashes: Arc<[TxHash]>,
 }
 
-/// See [`crate::common::Sequencer::is_ready`].
-#[derive(Debug, serde::Serialize)]
+/// See [`Sequencer::is_ready`].
+#[derive(Debug, Clone, serde::Serialize)]
 #[allow(missing_docs)]
 pub enum SequencerNotReadyDetails {
-    /// The node is catching up to the chain tip
+    /// The node is catching up to the chain tip.
     Syncing {
         target_da_height: u64,
         synced_da_height: u64,
     },
-    /// The sequencer is waiting for the DA to finalize more blocks
+    /// The sequencer is waiting for the DA to finalize more blocks.
     WaitingOnDa {
-        finalized_da_height: u64,
-        needed_finalized_height: u64,
-    },
-    /// The sequencer is waiting for the node to catch up to them
-    WaitingOnNode {
-        current_node_slot_number: u64,
-        current_sequencer_slot_number: u64,
-        current_delta: u64,
-        max_allowed_delta: u64,
+        finalized_slot_number: SlotNumber,
+        needed_finalized_slot_number: SlotNumber,
     },
     /// The sequencer is waiting for the blob sender to be ready.
     WaitingOnBlobSender {